refactor: 1개의 결과를 추출하는 DataAccessUtils.nullableSingleResult() 추가
