+++ jdbc/src/main/java/nextstep/dao/exception/EmptyResultDataAccessException.java
@@ -1,0 +1,8 @@
+package nextstep.dao.exception;
+
+public class EmptyResultDataAccessException extends DataAccessException {
+
+    public EmptyResultDataAccessException() {
+        super();
+    }
+}
