+++ jdbc/src/main/java/nextstep/dao/exception/IncorrectResultSizeDataAccessException.java
@@ -1,0 +1,8 @@
+package nextstep.dao.exception;
+
+public class IncorrectResultSizeDataAccessException extends DataAccessException {
+
+    public IncorrectResultSizeDataAccessException() {
+        super();
+    }
+}
