+++ jdbc/src/main/java/nextstep/jdbc/core/JdbcTemplate.java
@@ -6,6 +6,7 @@ import java.sql.ResultSet;
 import java.sql.SQLException;
 import java.util.List;
 import javax.sql.DataSource;
+import nextstep.dao.DataAccessUtils;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
@@ -36,14 +37,14 @@ public class JdbcTemplate {
     public <T> T queryForObject(final String sql, final RowMapper<T> rowMapper, final Object... args) {
         ResultSet rs = null;
         try (Connection conn = dataSource.getConnection();
-             PreparedStatement pstmt = conn.prepareStatement(sql)) {
+             PreparedStatement ps = conn.prepareStatement(sql)) {
             log.debug("query : {}", sql);
 
-            setPreparedStatement(pstmt, args);
-            rs = pstmt.executeQuery();
+            setPreparedStatement(ps, args);
+            rs = ps.executeQuery();
 
             RowMapperResultSetExtractor<T> rowMapperResultSetExtractor = new RowMapperResultSetExtractor<>(rowMapper);
-            return rowMapperResultSetExtractor.extractData(rs).get(0);
+            return DataAccessUtils.nullableSingleResult(rowMapperResultSetExtractor.extractData(rs));
         } catch (SQLException e) {
             log.error(e.getMessage(), e);
             throw new RuntimeException(e);
